+++ blackbook/src/config.ts
@@ -1,4 +1,4 @@
-import { Requester } from '@chainlink/external-adapter'
+import { Requester, logger } from '@chainlink/external-adapter'
 import { util } from '@chainlink/ea-bootstrap'
 import types from '@chainlink/types'
 import { NAME } from './endpoint/vehicle'
@@ -14,9 +14,23 @@ export type Config = types.Config & {
 }
 
 export const makeConfig = (prefix?: string): Config => {
+  const config = Requester.getDefaultConfig(prefix)
+  config.api.baseURL = config.api.baseURL || 'https://service.blackbookcloud.com/'
   return {
-    ...Requester.getDefaultConfig(prefix),
+    ...config,
     username: util.getRequiredEnv(ENV_API_USERNAME),
     password: util.getRequiredEnv(ENV_API_PASSWORD),
   }
+}
+
+// Config without sensitive data
+const redact = (config: Config) => ({
+  ...config,
+  apiKey: '*****',
+  username: '*****',
+  password: '*****',
+})
+
+export function logConfig(config: Config): void {
+  logger.debug('Adapter configuration:', { config: config && redact(config) })
 }