+++ marketcheck/src/endpoint/vehicle.ts
@@ -1,7 +1,5 @@
-import { AdapterError } from '@chainlink/external-adapter'
 import { Requester, Validator } from '@chainlink/external-adapter'
-import { ExecuteWithConfig } from '@chainlink/types'
-import { Config } from '../config'
+import { ExecuteWithConfig, Config } from '@chainlink/types'
 
 export const NAME = 'vehicle'
 
@@ -9,7 +7,6 @@ const customParams = {
   year: true,
   make: true,
   model: true,
-  customerid: false,
 }
 
 export const execute: ExecuteWithConfig<Config> = async (request, config) => {
@@ -17,26 +14,19 @@ export const execute: ExecuteWithConfig<Config> = async (request, config) => {
   if (validator.error) throw validator.error
 
   const jobRunID = validator.validated.id
+  const { year, make, model } = validator.validated.data
 
-  const { year, make, model, customerid } = validator.validated.data
+  const ymm = `${year}|${make}|${model}`
+
   const params = {
-    model,
-    customerid,
+    api_key: config.apiKey,
+    ymm,
   }
 
-  const auth = {
-    username: config.username,
-    password: config.password,
-  }
-
-  const reqConfig = {
-    url: `CPIAPI/CPIAPI/Vehicle/${year}/${make}`,
-    params,
-    auth,
-  }
-
+  const reqConfig = { ...config.api, params, url: 'sales/car' }
   const response = await Requester.request(reqConfig)
-  const path = ['cpi_vehicles', 'cpi_vehicle_list', 0, 'excellent']
+  // TODO: find path
+  const path = ['price']
   const result = Requester.validateResultNumber(response.data, path)
 
   return Requester.success(jobRunID, {